refactor(common): pull RCC clock select into the common directory
