+++ gantry/firmware/CMakeLists.txt
@@ -19,6 +19,8 @@ set(GANTRY_FW_NON_LINTABLE_SRCS
         ${COMMON_EXECUTABLE_DIR}/uart/uart.c
         ${COMMON_EXECUTABLE_DIR}/spi/spi.c
         ${COMMON_EXECUTABLE_DIR}/can/can.c
+        ${COMMON_EXECUTABLE_DIR}/errors/errors.c
+        ${COMMON_EXECUTABLE_DIR}/STM32G491RETx/clocking.c
         ${COMMON_EXECUTABLE_DIR}/motor_driver/motor.c)
 
 add_executable(gantry
